--- Assets/Editor/ShaderVariantCollectionDumpInfo.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace UTJ
-{
-    public class ShaderVariantCollectionDumpInfo
-    {
-        // Start is called before the first frame update
-        void Start()
-        {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-
-        }
-    }
-
-}
