+++ Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f0193bc706511f439b96da023826f9e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
