+++ Editor/AssetBundleChecker.Editor.asmdef
@@ -1,0 +1,14 @@
+{
+    "name": "AssetBundleChecker.Editor",
+    "references": [],
+    "optionalUnityReferences": [],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
+}
