+++ Editor/ShaderVariantCollectionDumpInfo.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace UTJ
+{
+    public class ShaderVariantCollectionDumpInfo
+    {
+    }
+
+}
