--- ProjectSettings/PresetManager.asset
@@ -1,6 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1386491679 &1
-PresetManager:
-  m_ObjectHideFlags: 0
-  m_DefaultList: []
