--- ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,0 @@
-m_EditorVersion: 2018.2.18f1
