+++ README.md.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: d72083d5c21e9ec44b8ecd4b145aa83e
-timeCreated: 1452438308
-licenseType: Pro
+guid: afc99a7c91afab74cbf97513cdf4a76b
 TextScriptImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
