+++ package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: c7cc24050e614ac429872a5750951ea5
-folderAsset: yes
-DefaultImporter:
+guid: 575abe9282a9d804ca953bbccbc6ccac
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
